Creating Register Style file
